chore(lint): fix lint issue with exported enum
