+++ .eslintrc.cjs
@@ -17,6 +17,6 @@ module.exports = {
   },
   plugins: ['react-refresh'],
   rules: {
-    'react-refresh/only-export-components': 'warn',
+    'react-refresh/only-export-components': 'error',
   },
 };
