+++ tsconfig.json
@@ -26,6 +26,7 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src", "custom.d.ts"],
+  // https://github.com/microsoft/TypeScript/issues/56749 Have to exclude *.mdx files
+  "include": ["custom.d.ts", "src/**/*.ts", "src/**/*.tsx"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
